Use 64-bit values to track frame times
This avoids overflow issues if the emulator is kept running
for a very long period of time (~49 days).

+++ include/cboy/gameboy.h
@@ -39,7 +39,7 @@ typedef struct gameboy {
     bool is_stopped, dma_requested;
 
     // to maintain the appropriate frame rate
-    uint32_t next_frame_time;
+    uint64_t next_frame_time;
 
     /* The Game Boy's internal 16-bit clock counter.
      * The DIV register at memory address 0xff04 is
